refactor: move utilities to lib/ and rename typeHelper to types
