+++ src/types.ts
@@ -1,7 +1,10 @@
-/**
- * Type that only includes properties that are possibly undefined.
- * @template T The type to process.
- */
-export type OnlyPossiblyUndefined<T> = {
-	[P in keyof T as undefined extends T[P] ? P : never]: T[P]
-}
+import type { AbstractTransformer } from './abstractTransformer'
+
+// biome-ignore lint/suspicious/noExplicitAny: using `any` for type parameter extraction - precision not required for conditional type inference
+export type InputOf<T> = T extends AbstractTransformer<infer I, any, any, any> ? I : never
+// biome-ignore lint/suspicious/noExplicitAny: using `any` for type parameter extraction - precision not required for conditional type inference
+export type OutputOf<T> = T extends AbstractTransformer<any, infer O, any, any> ? O : never
+// biome-ignore lint/suspicious/noExplicitAny: using `any` for type parameter extraction - precision not required for conditional type inference
+export type PropsOf<T> = T extends AbstractTransformer<any, any, infer P, any> ? P : never
+// biome-ignore lint/suspicious/noExplicitAny: using `any` for type parameter extraction - precision not required for conditional type inference
+export type IncludesOf<T> = T extends AbstractTransformer<any, any, any, infer Inc> ? Inc : never
