+++ tests/subTransformer.deep.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'bun:test'
 import { AbstractTransformer, AnyAbstractTransformer } from '../src/abstractTransformer'
-import { Cache } from '../src/cache'
+import { Cache } from '../src/lib/cache'
 
 // ─── Test Data Types ────────────────────────────────────────────────
